OutgoingMessageFactory introduced
